modif méthode validermodif

+++ Gallium v1/Vue/ModificationProduct.xaml.cs
@@ -58,7 +58,7 @@ namespace Gallium_v1.Vue
             {
                 this.produit.NomProduit = this.produitName.Text;
                 this.produit.Stock = Convert.ToInt32(this.stock.Text);
-                this.produit.PrixProduitAdhérent = Convert.ToInt32(this.prixAdherent.Text);
+                this.produit.PrixProduitAdhérent = Convert.ToDouble(this.prixAdherent.Text);
                 this.Close();
             }
         }
